Перенос сервисов в отдельную библеотеку
