--- Services/Webstore.Interfaces/Class1.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace Webstore.Interfaces
-{
-    public class Class1
-    {
-    }
-}
