--- Services/Webstore.Services/Class1.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace Webstore.Services
-{
-    public class Class1
-    {
-    }
-}
