+++ Services/Webstore.Services/Data/TestDB.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using WebStore.Domain.Entities;
 using WebStore.Domain.Models;
 
-namespace WebStore.Service
+namespace Webstore.Services.Data
 {
     public static class TestDB
     {
@@ -86,7 +86,7 @@ namespace WebStore.Service
                 Age = _rand.Next(20, 50)
             }).ToArray();
 
-        public static IEnumerable<BlogPost> Blogs { get; } = new []
+        public static IEnumerable<BlogPost> Blogs { get; } = new[]
         {
             new BlogPost
             {
@@ -130,7 +130,7 @@ namespace WebStore.Service
             },
         };
 
-        public static IEnumerable<Category> Сategories { get; } = new []
+        public static IEnumerable<Category> Сategories { get; } = new[]
         {
             new Category
             {
@@ -206,7 +206,7 @@ namespace WebStore.Service
             },
         };
 
-        public static IEnumerable<Brand> Brands { get; } = new []
+        public static IEnumerable<Brand> Brands { get; } = new[]
         {
             new Brand
             {
@@ -251,7 +251,7 @@ namespace WebStore.Service
             },
         };
 
-        public static IEnumerable<Product> Products { get; } = new []
+        public static IEnumerable<Product> Products { get; } = new[]
         {
             new Product
             {
@@ -259,8 +259,8 @@ namespace WebStore.Service
                 CategoryId=2,
                 BrandId=1,
                 Price=54490,
-                Name=@"Смартфон APPLE iPhone 11 64Gb, MWLT2RU/A, черный", 
-                ImageUrl="product_id1_1.jpg", 
+                Name=@"Смартфон APPLE iPhone 11 64Gb, MWLT2RU/A, черный",
+                ImageUrl="product_id1_1.jpg",
                 Description=@"ОС iPhone iOS 13, экран: 6.1"", IPS, 1792×828, процессор: Apple A13 Bionic, , GPS, ГЛОНАСС, с защитой от пыли и влаги, встроенная память: 64ГБ",
             },
             new Product
@@ -269,8 +269,8 @@ namespace WebStore.Service
                 CategoryId=2,
                 BrandId=2,
                 Price=20490,
-                Name=@"Смартфон SAMSUNG Galaxy A51 64Gb, SM-A515F, черный", 
-                ImageUrl="product_id2_1.jpg", 
+                Name=@"Смартфон SAMSUNG Galaxy A51 64Gb, SM-A515F, черный",
+                ImageUrl="product_id2_1.jpg",
                 Description=@"ОС Android 10, экран: 6.5"", Super AMOLED, 2400×1080, процессор: Exynos 9611, 2300МГц, 8-ми ядерный, GPS, ГЛОНАСС, оперативная память: 4ГБ, встроенная память: 64ГБ",
             },
             new Product
@@ -279,8 +279,8 @@ namespace WebStore.Service
                 CategoryId=2,
                 BrandId=3,
                 Price=22990,
-                Name=@"Смартфон XIAOMI Redmi Note 9 Pro 6/128Gb, зеленый", 
-                ImageUrl="product_id3_1.jpg", 
+                Name=@"Смартфон XIAOMI Redmi Note 9 Pro 6/128Gb, зеленый",
+                ImageUrl="product_id3_1.jpg",
                 Description=@"ОС Android 10.0, экран: 6.67"", IPS, 2400×1080, процессор: Qualcomm Snapdragon 720G, 2300МГц, 8-ми ядерный, FM-радио, GPS, ГЛОНАСС, время работы в режиме разговора, до: 33ч, оперативная память: 6ГБ, встроенная память: 128ГБ",
             },
             new Product
@@ -289,8 +289,8 @@ namespace WebStore.Service
                 CategoryId=2,
                 BrandId=4,
                 Price=12490,
-                Name=@"Смартфон ZTE Blade 20 Smart 128Gb, темный изумруд", 
-                ImageUrl="product_id4_1.jpg", 
+                Name=@"Смартфон ZTE Blade 20 Smart 128Gb, темный изумруд",
+                ImageUrl="product_id4_1.jpg",
                 Description=@"ОС Android 9.0, экран: 6.49"", IPS, 1560×720, процессор: MediaTek Helio P60, 2000МГц, 8-ми ядерный, FM-радио, GPS, ГЛОНАСС, время работы в режиме разговора, до: 50ч, в режиме ожидания, до: 833ч, оперативная память: 4ГБ, встроенная память: 128ГБ",
             },
             new Product
@@ -299,8 +299,8 @@ namespace WebStore.Service
                 CategoryId=2,
                 BrandId=5,
                 Price=10390,
-                Name=@"Смартфон HONOR 9A 64Gb, черный", 
-                ImageUrl="product_id5_1.jpg", 
+                Name=@"Смартфон HONOR 9A 64Gb, черный",
+                ImageUrl="product_id5_1.jpg",
                 Description=@"ОС Android 10 HMS, экран: 6.3"", IPS, 1600×720, процессор: MediaTek МТ6762R, 2000МГц, 8-ми ядерный, GPS, оперативная память: 3ГБ, встроенная память: 64ГБ",
             },
             new Product
@@ -309,8 +309,8 @@ namespace WebStore.Service
                 CategoryId=3,
                 BrandId=1,
                 Price=29900,
-                Name=@"Планшет APPLE iPad 2020 32Gb Wi-Fi MYL92RU/A, 32GB, iOS темно-серый", 
-                ImageUrl="product_id6_1.jpg", 
+                Name=@"Планшет APPLE iPad 2020 32Gb Wi-Fi MYL92RU/A, 32GB, iOS темно-серый",
+                ImageUrl="product_id6_1.jpg",
                 Description=@"сенсорный экран 10.2"" (25.9 см), разрешение: 2160 x 1620, Multitouch, Wi-Fi, Bluetooth, основная камера: 8Мп, фронтальная камера: 1.2Мп, fingerprint, встроенная память: 32ГБ, операционная система: iOS",
             },
             new Product
@@ -319,8 +319,8 @@ namespace WebStore.Service
                 CategoryId=3,
                 BrandId=2,
                 Price=25990,
-                Name=@"Планшет SAMSUNG Galaxy Tab S6 Lite SM-P610N, 4GB, 64GB, Android 10.0 серый [sm-p610nzaaser]", 
-                ImageUrl="product_id7_1.jpg", 
+                Name=@"Планшет SAMSUNG Galaxy Tab S6 Lite SM-P610N, 4GB, 64GB, Android 10.0 серый [sm-p610nzaaser]",
+                ImageUrl="product_id7_1.jpg",
                 Description=@"сенсорный экран 10.4"" (26.4 см), разрешение: 2000 x 1200, Multitouch, Wi-Fi, Bluetooth, основная камера: 8Мп, фронтальная камера: 5Мп, встроенная память: 64ГБ, операционная система: Android 10.0",
             },
             new Product
@@ -329,8 +329,8 @@ namespace WebStore.Service
                 CategoryId=3,
                 BrandId=5,
                 Price=28990,
-                Name=@"Планшет HONOR Pad V6, 6ГБ, 128GB, Android 10.0 черный [53011eta]", 
-                ImageUrl="product_id8_1.jpg", 
+                Name=@"Планшет HONOR Pad V6, 6ГБ, 128GB, Android 10.0 черный [53011eta]",
+                ImageUrl="product_id8_1.jpg",
                 Description=@"сенсорный экран 10.4"" (26.4 см), разрешение: 2000 x 1200, Multitouch, Wi-Fi, Bluetooth, GPS, ГЛОНАСС, основная камера: 13Мп, фронтальная камера: 8Мп, встроенная память: 128ГБ, операционная система: Android 10.0",
             },
             new Product
@@ -339,8 +339,8 @@ namespace WebStore.Service
                 CategoryId=18,
                 BrandId=3,
                 Price=2120,
-                Name=@"Фитнес-трекер XIAOMI Band 4 XMSH07HM, 0.95"", черный / черный [mgw4057ru]", 
-                ImageUrl="product_id9_1.jpg", 
+                Name=@"Фитнес-трекер XIAOMI Band 4 XMSH07HM, 0.95"", черный / черный [mgw4057ru]",
+                ImageUrl="product_id9_1.jpg",
                 Description=@"Mi Band 4 — очень удачный фитнес-браслет, который оказался улучшен практически во всех аспектах в сравнении с предшественниками. Устройство не только получило великолепный AMOLED-дисплей, но и оказалось оборудовано продвинутыми датчиками.",
             },
             new Product
@@ -349,8 +349,8 @@ namespace WebStore.Service
                 CategoryId=5,
                 BrandId=1,
                 Price=12990,
-                Name=@"Гарнитура APPLE AirPods, with Charging Case, Bluetooth, вкладыши, белый [mv7n2ru/a]", 
-                ImageUrl="product_id10_1.jpg", 
+                Name=@"Гарнитура APPLE AirPods, with Charging Case, Bluetooth, вкладыши, белый [mv7n2ru/a]",
+                ImageUrl="product_id10_1.jpg",
                 Description=@"тип соединения: беспроводные bluetooth; акустический тип: закрытые; тип амбушюр: вкладыши",
             },
             new Product
@@ -359,8 +359,8 @@ namespace WebStore.Service
                 CategoryId=7,
                 BrandId=1,
                 Price=70990,
-                Name=@"Ноутбук APPLE MacBook Air 13.3"", Intel Core i5 5350U 1.8ГГц, 8ГБ, 128ГБ SSD, Intel HD Graphics 6000, Mac OS X El Capitan, MQD32RU/A, серебристый", 
-                ImageUrl="product_id11_1.jpg", 
+                Name=@"Ноутбук APPLE MacBook Air 13.3"", Intel Core i5 5350U 1.8ГГц, 8ГБ, 128ГБ SSD, Intel HD Graphics 6000, Mac OS X El Capitan, MQD32RU/A, серебристый",
+                ImageUrl="product_id11_1.jpg",
                 Description=@"экран: 13.3""; разрешение экрана: 1440×900; процессор: Intel Core i5 5350U; частота: 1.8 ГГц (2.9 ГГц, в режиме Turbo); память: 8192 Мб, LPDDR3, 1600 МГц; SSD: 128 Гб; Intel HD Graphics 6000; WiFi; Bluetooth; WEB-камера; Mac OS X El Capitan",
             },
             new Product
@@ -369,8 +369,8 @@ namespace WebStore.Service
                 CategoryId=7,
                 BrandId=5,
                 Price=5990,
-                Name=@"Ультрабук HONOR MagicBook Pro HLY-W19R, 16.1"", IPS, AMD Ryzen 5 3550H 2.1ГГц, 8ГБ, 512ГБ SSD, AMD Radeon Vega 8, Windows 10, 53010TSA, серый", 
-                ImageUrl="product_id12_1.jpg", 
+                Name=@"Ультрабук HONOR MagicBook Pro HLY-W19R, 16.1"", IPS, AMD Ryzen 5 3550H 2.1ГГц, 8ГБ, 512ГБ SSD, AMD Radeon Vega 8, Windows 10, 53010TSA, серый",
+                ImageUrl="product_id12_1.jpg",
                 Description=@"экран: 16.1""; разрешение экрана: 1920×1080; тип матрицы: IPS; процессор: AMD Ryzen 5 3550H; частота: 2.1 ГГц (3.7 ГГц, в режиме Turbo); память: 8192 Мб, DDR4, 2400 МГц; SSD: 512 Гб; AMD Radeon Vega 8; WiFi; Bluetooth; HDMI; WEB-камера; Windows 10",
             },
             new Product
@@ -379,8 +379,8 @@ namespace WebStore.Service
                 CategoryId=8,
                 BrandId=1,
                 Price=118890,
-                Name=@"Моноблок APPLE iMac MRT42RU/A, 21.5"", Intel Core i5 8500, 8ГБ, 1000ГБ, AMD Radeon Pro 560X - 4096 Мб, Mac OS, серебристый и черный", 
-                ImageUrl="product_id13_1.jpg", 
+                Name=@"Моноблок APPLE iMac MRT42RU/A, 21.5"", Intel Core i5 8500, 8ГБ, 1000ГБ, AMD Radeon Pro 560X - 4096 Мб, Mac OS, серебристый и черный",
+                ImageUrl="product_id13_1.jpg",
                 Description=@"экран 21.5"", 4096 х 2304; процессор: Intel Core i5 8500, 3.0 ГГц (4.1 ГГц, в режиме Turbo); оперативная память: DDR4 8192 Мб 2666 МГц; видеокарта: AMD Radeon Pro 560X — 4096 Мб; HDD: 1000 Гб; Web-камера; Wi-Fi; Bluetooth",
             },
             new Product
@@ -389,8 +389,8 @@ namespace WebStore.Service
                 CategoryId=10,
                 BrandId=6,
                 Price=10790,
-                Name=@"Процессор AMD Ryzen 5 2600, SocketAM4, OEM [yd2600bbm6iaf]", 
-                ImageUrl="product_id14_1.jpg", 
+                Name=@"Процессор AMD Ryzen 5 2600, SocketAM4, OEM [yd2600bbm6iaf]",
+                ImageUrl="product_id14_1.jpg",
                 Description=@"сокет SocketAM4, ядро Pinnacle Ridge, ядер — 6, потоков — 12, частота 3.4 ГГц и 3.9 ГГц в режиме Turbo, L3 кэш 16МБ, техпроцесс 12нм, поддержка памяти DDR4 каналов памяти — 2, множитель не заблокирован, контроллер PCI Express 3.0, поставка OEM",
             },
             new Product
@@ -399,8 +399,8 @@ namespace WebStore.Service
                 CategoryId=10,
                 BrandId=7,
                 Price=12390,
-                Name=@"Процессор INTEL Core i5 9400F, LGA 1151v2, OEM [cm8068403358819s rf6m]", 
-                ImageUrl="product_id15_1.jpg", 
+                Name=@"Процессор INTEL Core i5 9400F, LGA 1151v2, OEM [cm8068403358819s rf6m]",
+                ImageUrl="product_id15_1.jpg",
                 Description=@"сокет LGA 1151v2, ядро Coffee Lake, ядер — 6, потоков — 6, частота 2.9 ГГц и 4.1 ГГц в режиме Turbo, L3 кэш 9МБ, техпроцесс 14нм, поддержка памяти DDR4 до 128 ГБ, каналов памяти — 2, контроллер PCI Express 3.0, поставка OEM",
             },
             new Product
@@ -409,8 +409,8 @@ namespace WebStore.Service
                 CategoryId=11,
                 BrandId=8,
                 Price=5290,
-                Name=@"Материнская плата GIGABYTE B450M S2H, SocketAM4, AMD B450, mATX, Ret", 
-                ImageUrl="product_id16_1.jpg", 
+                Name=@"Материнская плата GIGABYTE B450M S2H, SocketAM4, AMD B450, mATX, Ret",
+                ImageUrl="product_id16_1.jpg",
                 Description=@"гнездо процессора: SocketAM4; чипсет AMD B450; память DDR4 — 2слота; частотой до 2933МГц; SATA RAID, тип поставки: Ret; mATX",
             },
             new Product
@@ -419,8 +419,8 @@ namespace WebStore.Service
                 CategoryId=11,
                 BrandId=9,
                 Price=4690,
-                Name=@"Материнская плата ASUS PRIME H310M-R R2.0, LGA 1151v2, Intel H310, mATX, Ret (White Box)", 
-                ImageUrl="product_id17_1.jpg", 
+                Name=@"Материнская плата ASUS PRIME H310M-R R2.0, LGA 1151v2, Intel H310, mATX, Ret (White Box)",
+                ImageUrl="product_id17_1.jpg",
                 Description=@"гнездо процессора: LGA 1151v2; чипсет Intel H310; память DDR4 — 2слота; частотой до 2666МГц; тип поставки: Ret (White Box); mATX",
             },
             new Product
@@ -429,8 +429,8 @@ namespace WebStore.Service
                 CategoryId=12,
                 BrandId=8,
                 Price=21400,
-                Name=@"Видеокарта GIGABYTE nVidia GeForce GTX 1660 , GV-N1660OC-6GD, 6ГБ, GDDR5, OC, Ret", 
-                ImageUrl="product_id18_1.jpg", 
+                Name=@"Видеокарта GIGABYTE nVidia GeForce GTX 1660 , GV-N1660OC-6GD, 6ГБ, GDDR5, OC, Ret",
+                ImageUrl="product_id18_1.jpg",
                 Description=@"nVidia GeForce GTX 1660 ; частота процессора: 1785 МГц (1830 МГц, в режиме Boost); частота памяти: 8002МГц; объём видеопамяти: 6ГБ; тип видеопамяти: GDDR5; OverClock Edition; DirectX 12/OpenGL 4.5; доп. питание: 8 pin; блок питания не менее: 450Вт; тип поставки: Ret",
             },
             new Product
@@ -439,8 +439,8 @@ namespace WebStore.Service
                 CategoryId=13,
                 BrandId=8,
                 Price=7740,
-                Name=@"Блок питания GIGABYTE GP-P750GM, 750Вт, 120мм, черный, retail [28200-p750g-1eur]", 
-                ImageUrl="product_id19_1.jpg", 
+                Name=@"Блок питания GIGABYTE GP-P750GM, 750Вт, 120мм, черный, retail [28200-p750g-1eur]",
+                ImageUrl="product_id19_1.jpg",
                 Description=@"размер вентилятора 120мм; ATX; мощность: 750Вт; активный PFC; стандарт 80 PLUS GOLD; питание MB и CPU: 24+2x(4+4) pin; питание видеокарты: 4х(6+2) pin; разъемы Molex: 3шт; разъемы SATA: 8шт; отсоединяющиеся кабели; цвет: черный; тип поставки Ret",
             },
             new Product
@@ -449,8 +449,8 @@ namespace WebStore.Service
                 CategoryId=15,
                 BrandId=3,
                 Price=23990,
-                Name=@"Телевизор XIAOMI Mi TV 4S 43, 43"", Ultra HD 4K", 
-                ImageUrl="product_id20_1.jpg", 
+                Name=@"Телевизор XIAOMI Mi TV 4S 43, 43"", Ultra HD 4K",
+                ImageUrl="product_id20_1.jpg",
                 Description=@"диагональ: 43""; разрешение: 3840 x 2160; Ultra HD 4K; SMART TV; яркость: 220кд/м2; DVB-T2; DVB-С; разъем Ethernet 1; VESA 300×300; USB: 2: HDMI: 3 : цвет: черный",
             },
             new Product
@@ -459,8 +459,8 @@ namespace WebStore.Service
                 CategoryId=16,
                 BrandId=3,
                 Price=5700,
-                Name=@"Медиаплеер XIAOMI Mi TV Box S", 
-                ImageUrl="product_id21_1.jpg", 
+                Name=@"Медиаплеер XIAOMI Mi TV Box S",
+                ImageUrl="product_id21_1.jpg",
                 Description=@"поддержка HDTV 4096p; Wi-Fi 802.11ac; USB 2.0: 1",
             },
             new Product
@@ -469,8 +469,8 @@ namespace WebStore.Service
                 CategoryId=17,
                 BrandId=10,
                 Price=4540,
-                Name=@"Колонки CREATIVE T3250W, 2.1, черный [51mf0450aa000]", 
-                ImageUrl="product_id22_1.jpg", 
+                Name=@"Колонки CREATIVE T3250W, 2.1, черный [51mf0450aa000]",
+                ImageUrl="product_id22_1.jpg",
                 Description=@"акустический тип: 2.1; выходная мощность (RMS) 9Вт; регуляторы на пульте; питание от сети; тип акустики: стационарная; цвет: черный",
             },
         };
