+++ Services/Webstore.Services/Products/InMemoryEmployeesData.cs
@@ -2,11 +2,11 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Webstore.Interfaces.Services;
 using WebStore.Domain.Models;
-using WebStore.Infrastructure.Interfaces;
-using WebStore.Service;
+using Webstore.Services.Data;
 
-namespace WebStore.Infrastructure.Services
+namespace Webstore.Services.Products
 {
     public class InMemoryEmployeesData : IEmployeesData
     {
