+++ Services/Webstore.Services/Webstore.Services.csproj
@@ -1,7 +1,13 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
+    <TargetFramework>netstandard2.1</TargetFramework>
   </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\Common\WebStore.Domain\WebStore.Domain.csproj" />
+    <ProjectReference Include="..\Webstore.DAL\Webstore.DAL.csproj" />
+    <ProjectReference Include="..\Webstore.Interfaces\Webstore.Interfaces.csproj" />
+  </ItemGroup>
 
 </Project>
