+++ UI/WebStore/Controllers/BlogController.cs
@@ -4,8 +4,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using WebStore.Domain.Models;
-using WebStore.Infrastructure.Interfaces;
 using WebStore.Domain.ViewModels;
+using Webstore.Interfaces.Services;
 
 namespace WebStore.Controllers
 {
