+++ UI/WebStore/Controllers/CartController.cs
@@ -4,8 +4,8 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Webstore.Interfaces.Services;
 using WebStore.Domain.ViewModels;
-using WebStore.Infrastructure.Interfaces;
 
 namespace WebStore.Controllers
 {
