+++ UI/WebStore/Controllers/ShopController.cs
@@ -3,11 +3,11 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using WebStore.Domain;
-using WebStore.Infrastructure.Interfaces;
 using WebStore.Domain.ViewModels;
 using WebStore.Domain.Entities;
-using WebStore.Infrastructure.Mapping;
 using WebStore.Domain.ViewModels;
+using Webstore.Interfaces.Services;
+using Webstore.Services.Mapping;
 
 namespace WebStore.Controllers
 {
