+++ UI/WebStore/Startup.cs
@@ -10,15 +10,13 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Webstore.DAL.Context;
-using WebStore.Data;
-using WebStore.Infrastructure.Interfaces;
-using WebStore.Infrastructure.Services;
-using WebStore.Service;
-using WebStore.Infrastructure.Services.InSql;
+using Webstore.Services.Products;
 using WebStore.Domain.Entities.Identity;
 using Microsoft.AspNetCore.Identity;
-using WebStore.Infrastructure.Services.InCookies;
-using WebStore.Infrastructure.Services.InSQL;
+using Webstore.Interfaces.Services;
+using Webstore.Services.Data;
+using Webstore.Services.Products.InCookies;
+using Webstore.Services.Products.InSql;
 
 namespace WebStore
 {
