+++ UI/WebStore/WebStore.csproj
@@ -22,6 +22,8 @@
   <ItemGroup>
     <ProjectReference Include="..\..\Common\WebStore.Domain\WebStore.Domain.csproj" />
     <ProjectReference Include="..\..\Services\Webstore.DAL\Webstore.DAL.csproj" />
+    <ProjectReference Include="..\..\Services\Webstore.Interfaces\Webstore.Interfaces.csproj" />
+    <ProjectReference Include="..\..\Services\Webstore.Services\Webstore.Services.csproj" />
   </ItemGroup>
 
   <ProjectExtensions><VisualStudio><UserProperties bundleconfig_1json__JsonSchema="" /></VisualStudio></ProjectExtensions>
